+++ Modules/Microsoft365DSC/DSCResources/MSFT_IntuneDiskEncryptionFileVaultPolicyMacOS/MSFT_IntuneDiskEncryptionFileVaultPolicyMacOS.psm1
@@ -329,12 +329,13 @@ function Set-TargetResource
             -TemplateId $templateReferenceId
 
         $createParameters = @{
-            Name              = $DisplayName
-            Description       = $Description
-            TemplateReference = @{ templateId = $templateReferenceId }
-            Platforms         = $platforms
-            Technologies      = $technologies
-            Settings          = $settings
+            name              = $DisplayName
+            description       = $Description
+            templateReference = @{ templateId = $templateReferenceId }
+            platforms         = $platforms
+            technologies      = $technologies
+            settings          = $settings
+            roleScopeTagIds   = $RoleScopeTagIds
         }
 
         #region resource generator code