+++ Modules/Microsoft365DSC/DSCResources/MSFT_IntuneSettingCatalogCustomPolicyWindows10/MSFT_IntuneSettingCatalogCustomPolicyWindows10.psm1
@@ -114,7 +114,7 @@ function Get-TargetResource
 
             if (-not [string]::IsNullOrEmpty($Name))
             {
-                $getValue = Get-MgBetaDeviceManagementConfigurationPolicy `
+                [array]$getValue = Get-MgBetaDeviceManagementConfigurationPolicy `
                     -Filter "Name eq '$($Name -replace "'", "''")' and Platforms eq 'windows10' and Technologies eq 'mdm' and TemplateReference/TemplateFamily eq 'none'" `
                     -All `
                     -ErrorAction SilentlyContinue